requirements-dashboard.txt
flask>=3.0
python-dotenv>=1.0
alpaca-py>=0.21
requests>=2.31.0
